requirements.txt
flask==1.1.2
flask-restful==0.3.8
flask-jwt-extended==4.1.0
flask-socketio==5.0.1
python-dotenv==0.15.0
flask-mongoengine==1.0.0
pytest==6.2.2
flake8==3.9.0
flask-cors==3.0.10

wapi-python
pytz~=2021.1
entsoe-py
mongoengine~=0.23.0
Werkzeug~=1.0.1
pandas~=1.1.5